+++ common/tests/unittests/test_common.py
@@ -1,4 +1,5 @@
 import autogluon.common
 
+
 def test_common():
 	print("Not Implemented")
