+++ common/tests/unittests/test_import_version.py
@@ -1,0 +1,6 @@
+import autogluon.common
+
+
+def test_import_version():
+    assert isinstance(autogluon.common.__version__, str)
+    assert len(autogluon.common.__version__) != 0
