+++ core/src/autogluon/core/augmentation/distill_utils.py
@@ -5,10 +5,11 @@ import numpy as np
 import pandas as pd
 from sklearn.neighbors import NearestNeighbors
 
+from autogluon.common.features.feature_metadata import FeatureMetadata
+from autogluon.common.features.types import R_CATEGORY, R_FLOAT, R_INT
+
 from ..constants import BINARY, MULTICLASS, REGRESSION
 from ..metrics import mean_squared_error
-from ..features.feature_metadata import FeatureMetadata
-from ..features.types import R_CATEGORY, R_FLOAT, R_INT
 
 logger = logging.getLogger(__name__)
