+++ features/src/autogluon/features/generators/text_special.py
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 from pandas import DataFrame, Series
 
-from autogluon.core.features.types import S_IMAGE_PATH, S_TEXT, S_TEXT_SPECIAL
+from autogluon.common.features.types import S_IMAGE_PATH, S_TEXT, S_TEXT_SPECIAL
 
 from .abstract import AbstractFeatureGenerator
 from .binned import BinnedFeatureGenerator
